backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
geopy==2.3.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
Werkzeug==2.3.7
SQLAlchemy==2.0.20
GeoAlchemy2==0.13.3
Shapely==2.0.1 